added component templates
- v8: reduce dynamtic template with path_match
- msg: add path unmatch
- es: parse date for ES
- rename component templates & sort template JSON keys
- test: try installing component templates
