+++ config/elasticsearch.ini
@@ -20,13 +20,13 @@
 [cloud]
 ; id=PROJECT-NAME:Long-base64-string
 
-; [auth]
+[auth]
 ; username=haraka
 ; password=nice-long-pass-phrase
 ; apiKey=base64-string
 ; bearer=token
 
-; [tls]
+[tls]
 ; rejectUnauthorized=false
