--- templates/component/dkim.json
@@ -1,45 +1,0 @@
-{
-	"component_templates": [
-		{
-			"name": "haraka-plugin-dkim",
-			"component_template": {
-				"template": {
-					"mappings": {
-						"properties": {
-							"p.dkim.domain": {
-								"eager_global_ordinals": false,
-								"index_phrases": false,
-								"fielddata": false,
-								"norms": true,
-								"index": true,
-								"store": false,
-								"type": "text",
-								"index_options": "positions"
-							},
-							"p.dkim.identity": {
-								"eager_global_ordinals": false,
-								"index_phrases": false,
-								"fielddata": false,
-								"norms": true,
-								"index": true,
-								"store": false,
-								"type": "text",
-								"index_options": "positions"
-							},
-							"p.dkim.selector": {
-								"eager_global_ordinals": false,
-								"norms": false,
-								"index": true,
-								"store": false,
-								"type": "keyword",
-								"split_queries_on_whitespace": false,
-								"index_options": "docs",
-								"doc_values": true
-							}
-						}
-					}
-				}
-			}
-		}
-	]
-}
