+++ templates/component/plugin-bounce.json
@@ -1,0 +1,22 @@
+{
+  "component_templates": [
+    {
+      "component_template": {
+        "template": {
+          "mappings": {
+            "properties": {
+              "p.bounce.isa": {
+                "doc_values": true,
+                "index": true,
+                "store": false,
+                "type": "boolean"
+              }
+            }
+          }
+        },
+        "version": 1
+      },
+      "name": "haraka-plugin-bounce"
+    }
+  ]
+}
