+++ templates/component/plugin-rspamd.json
@@ -1,0 +1,41 @@
+{
+  "component_templates": [
+    {
+      "component_template": {
+        "template": {
+          "mappings": {
+            "properties": {
+              "p.rspamd.action": {
+                "type": "keyword"
+              },
+              "p.rspamd.emails": {
+                "eager_global_ordinals": false,
+                "fielddata": false,
+                "index": true,
+                "index_options": "positions",
+                "index_phrases": false,
+                "norms": true,
+                "store": false,
+                "type": "text"
+              },
+              "p.rspamd.required_score": {
+                "type": "float"
+              },
+              "p.rspamd.score": {
+                "type": "float"
+              },
+              "p.rspamd.time": {
+                "type": "float"
+              },
+              "p.rspamd.urls": {
+                "type": "text"
+              }
+            }
+          }
+        },
+        "version": 1
+      },
+      "name": "haraka-plugin-rspamd"
+    }
+  ]
+}
