+++ templates/component/plugin-spf.json
@@ -1,0 +1,25 @@
+{
+  "component_templates": [
+    {
+      "component_template": {
+        "template": {
+          "mappings": {
+            "properties": {
+              "p.spf.domain": {
+                "type": "keyword"
+              },
+              "p.spf.result": {
+                "type": "keyword"
+              },
+              "p.spf.scope": {
+                "type": "keyword"
+              }
+            }
+          }
+        },
+        "version": 1
+      },
+      "name": "haraka-plugin-spf"
+    }
+  ]
+}
