+++ templates/component/results.json
@@ -1,0 +1,62 @@
+{
+  "component_templates": [
+    {
+      "component_template": {
+        "template": {
+          "mappings": {
+            "dynamic_templates": [
+              {
+                "fail_results": {
+                  "mapping": {
+                    "index": "true",
+                    "type": "keyword"
+                  },
+                  "path_match": "*.fail"
+                }
+              },
+              {
+                "pass_results": {
+                  "mapping": {
+                    "index": "true",
+                    "type": "keyword"
+                  },
+                  "path_match": "*.pass"
+                }
+              },
+              {
+                "skip_results": {
+                  "mapping": {
+                    "index": "true",
+                    "type": "keyword"
+                  },
+                  "path_match": "*.skip"
+                }
+              },
+              {
+                "msg_results": {
+                  "mapping": {
+                    "index": "true",
+                    "type": "keyword"
+                  },
+                  "path_match": "*.msg",
+                  "path_unmatch": "*.count.msg"
+                }
+              },
+              {
+                "err_results": {
+                  "mapping": {
+                    "index": "true",
+                    "type": "keyword"
+                  },
+                  "path_match": "*.err"
+                }
+              }
+            ]
+          }
+        },
+        "version": 1
+      },
+      "name": "haraka-results"
+    }
+  ]
+}
