+++ test/index.js
@@ -207,78 +207,3 @@ describe('trim_plugin_name', function () {
     done()
   })
 })
-
-describe('storesIndexMapTemplate', function () {
-  beforeEach(setup)
-
-  it('saves an index map template to Elasticsearch', function (done) {
-    this.timeout(4000)
-
-    const plugin = this.plugin
-    const filePath = path.resolve('templates', 'index', 'v8.json')
-
-    plugin.load_es_ini()
-
-    plugin.es_connect((err) => {
-      assert.ifError(err)
-
-      if (err) {
-        done()
-        return
-      }
-
-      fs.readFile(filePath, 'utf8', (err2, data) => {
-        if (err2) {
-          console.error(err2)
-          done()
-        }
-
-        plugin.es.indices
-          .putTemplate({
-            name: 'haraka-results',
-            index_patterns: 'smtp-*',
-            mappings: JSON.parse(data),
-          })
-          .then((result) => {
-            console.log(result)
-          })
-          .catch((err3) => {
-            if (err3.status !== 404) {
-              console.error(err3)
-            }
-            // other tests are running, so currently
-            // stored mapping may conflict
-          })
-          .finally(done)
-      })
-    })
-  })
-})
-
-describe('log_connection', function () {
-  beforeEach(setup)
-
-  it('saves results to Elasticsearch', function (done) {
-    const plugin = this.plugin
-
-    plugin.load_es_ini()
-    plugin.es_connect(function (err) {
-      assert.ifError(err)
-
-      console.log('giving ES a few secs to start up')
-
-      const connection = fixtures.connection.createConnection()
-      connection.local.ip = '127.0.0.1'
-      connection.remote.ip = '172.1.1.1'
-      connection.uuid = utils.uuid()
-      connection.count = { msg: { accepted: 1 } }
-      connection.results.add({ name: 'rspamd' }, { msg: 'test' })
-
-      // console.log(util.inspect(connection, { depth: null }));
-      plugin.log_connection(function () {
-        // assert.ok(1);
-        done()
-      }, connection)
-    })
-  })
-})
